# api-docs/swagger.yaml
swagger: '2.0'
info:
  description: 'API Evangelion Anime'
  version: '1.0.0'
  title: 'API del blog acerca de Evangelion'
host: 'localhost:5000'
basePath: '/'
schemes:
  - 'http'
paths:
  /posts:
    get:
      summary: 'Obtiene todos los posts publicados hasta el momento mostrando el título, contenido, autor, categoria y tags del post.'
      responses:
        200:
          description: 'Una lista de posts creados acerca de diversos temas del anime de Evangelion.'
          schema:
            type: 'array'
            items:
              $ref: '#/definitions/Post'
    post:
      summary: 'Crea un nuevo post, se tendrán que ingresar el título, contenido, autor, categoria y tags del post.'
      consumes:
        - 'application/json'
      parameters:
        - in: 'body'
          name: 'body'
          description: 'Elementos para crear posts dentro de la API tal como: titulo, contenido, autor, categoria y tags.'
          required: true
          schema:
            $ref: '#/definitions/Post'
      responses:
        200:
          description: 'Post creado exitosamente.'
        500:
          description: 'Ha ocurrido un error.'

  /posts/{postId}:
    get:
      summary: 'Obtiene un post específico por ID.'
      parameters:
        - in: 'path'
          name: 'postId'
          required: true
          type: 'integer'
          description: 'El ID del post a obtener'
      responses:
        200:
          description: 'Detalles del post solicitado.'
          schema:
            $ref: '#/definitions/Post'
        404:
          description: 'Post no encontrado.'

        500:
          description: 'Ha ocurrido un error.'

    put:
      summary: 'Actualiza un post existente.'
      consumes:
        - 'application/json'
      parameters:
        - in: 'path'
          name: 'postId'
          required: true
          type: 'integer'
          description: 'El ID del post a actualizar'
        - in: 'body'
          name: 'body'
          description: 'Información a actualizar del post'
          required: true
          schema:
            $ref: '#/definitions/Post'
      responses:
        200:
          description: 'Post actualizado exitosamente.'
        404:
          description: 'Post no encontrado.'

        500:
          description: 'Ha ocurrido un error.'

    delete:
      summary: 'Elimina un post específico.'
      parameters:
        - in: 'path'
          name: 'postId'
          required: true
          type: 'integer'
          description: 'El ID del post a eliminar'
      responses:
        200:
          description: 'Post eliminado exitosamente.'
        404:
          description: 'Post no encontrado.'

        500:
          description: 'Ha ocurrido un error.'

definitions:
  Post:
    type: 'object'
    properties:
      title:
        type: 'string'
      content:
        type: 'string'
      author:
        type: 'string'
      category:
        type: 'string'
      tags:
        type: 'string'